app/pyproject.toml
[tool.poetry]
name = "app"
version = "0.1.0"
description = ""
authors = ["HarlemShaker <ega3001@yandex.ru>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
fastapi = "^0.111.1"
sqlalchemy = "^2.0.31"
alembic = "^1.13.2"
attrs = "^23.2.0"
environ-config = "^23.2.0"
asyncpg = "^0.29.0"
psycopg2 = "^2.9.9"
gunicorn = "^22.0.0"
pydantic = "^2.8.2"

[tool.poetry.group.dev.dependencies]
mypy = "^1.11.0"
black = "^24.4.2"
pytest = "^8.3.2"

[tool.black]
line-length = 88
target-version = ['py39']
include = '\.pyi?$'

[tool.mypy]
ignore_missing_imports = true
show_column_numbers = true
disallow_untyped_defs = true
disallow_untyped_calls = true

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
